requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.3.1
apimatic-core==0.2.16
apimatic-core-interfaces==0.1.5
apimatic-requests-client-adapter==0.1.6
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
billiard==4.2.0
boto==2.49.0
boto3==1.35.4
botocore==1.35.4
build==1.2.1
CacheControl==0.12.14
cachetools==5.5.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.3
cryptography==43.0.0
defusedxml==0.7.1
Deprecated==1.2.14
deprecation==2.1.0
Django==5.1.3
django-admin-honeypot==1.1.0
django-allauth==64.1.0
django-appconf==1.0.6
django-bootstrap4==24.3
django-compressor==4.5.1
django-cors-headers==4.4.0
django-htmlmin==0.11.0
django-markdownx==4.0.7
django-ratelimit==4.1.0
django-redis==5.4.0
django-storages==1.14.4
django-timezone-field==7.0
django_csp==3.8
docutils==0.21.2
frozenlist==1.4.1
git-filter-repo==2.45.0
gnupg==2.3.1
google-api-core==2.19.1
google-api-python-client==2.142.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.63.2
gunicorn==23.0.0
html5lib==1.1
httplib2==0.22.0
idna==3.7
itsdangerous==2.2.0
jmespath==1.0.1
jsonpickle==3.0.4
jsonpointer==2.4
kombu==5.4.0
lxml==5.3.0
Markdown==3.7
msgpack==1.0.8
multidict==6.0.5
mysql-connector-python==9.0.0
mysqlclient==2.2.4
oauthlib==3.2.2
packaging==24.1
pillow==10.4.0
pip-tools==7.4.1
pipdeptree==2.23.1
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==5.27.3
psutil==6.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.12.1
pycparser==2.22
PyJWT==2.9.0
pymemcache==4.0.0
pyOpenSSL==24.2.1
pyparsing==3.1.2
PyPDF2==3.0.1
pypng==0.20220715.0
pyproject_hooks==1.1.0
python-crontab==3.2.0
python-dateutil==2.8.2
python-decouple==3.8
python-memcached==1.62
python3-openid==3.2.0
pytz==2024.1
rcssmin==1.1.2
redis==5.0.8
requests==2.32.3
requests-oauthlib==2.0.0
rjsmin==1.2.2
rsa==4.9
s3transfer==0.10.4
setuptools==75.6.0
six==1.16.0
soupsieve==2.6
sqlparse==0.5.2
squareup==38.0.0.20240821
suit==2.0.2
toml==0.10.2
twilio==9.2.3
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
wheel==0.44.0
whitenoise==6.7.0
wrapt==1.16.0
yarl==1.9.4
